+++ person.rb
@@ -1,0 +1,28 @@
+class Person
+  def initialize(age, name = 'Unknown', parent_permission: true)
+    @id = Random.rand(1..1000)
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
+  end
+
+  def _id
+    @id
+  end
+
+  def _name
+    @name
+  end
+
+  def _age
+    @age
+  end
+
+  def of_age?
+    @age >= 18
+  end
+
+  def can_use_services?
+    @age >= 18 || @parent_permission
+  end
+end
